Add common Upbound issue and PR automation
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file

This commit adds global issue and PR templates per the above link. If I follow
correctly these templates will become the default for any repos within the
upbound org that don't have their own templates configured. If this PR is merged
I'd like to do a pass through upbound repos and delete repo specific templates
if they're not an improvement over these common ones.

https://docs.github.com/en/actions/learn-github-actions/sharing-workflows-with-your-organization
https://github.com/marketplace/actions/global-workflows-support

This commit also adds a broadly useful GitHub action that can be used to create
a tag. As it stands we might need to use the 'global workflows' action to enable
this, since GitHub appears to only support workflow _templates_ natively.

I considered adding more checklists to the issue and PR templates included in
this PR, but it seems like we could automate most of the things we'd prompt the
issue or PR author to handle. For example:

* Linking issues to an epic using an 'Epic: upbound/example#23' style comment.
* Adding a size estimate label to issues and aggregating size at the sprint
  column and/or epic level.
* Adding a size assessment to PRs (e.g. size/s, size/xl).
* Automatically backporting PRs to release branches.

Signed-off-by: Nic Cope <negz@rk0n.org>