Add label prompts for size and epic links
The idea is that these labels will prompt folks to replace the labels with a
link to an epic and a size estimate. I prefer labels here to a checklist in the
issue because not all contributors will have the context and access needed to
create labels and link to epics.

Signed-off-by: Nic Cope <negz@rk0n.org>