Restrict who many issue commands
This commit updates the Comment Commands job to use an action that handles
comment commands, because said action is able to ensure that the job will fail
if the commenter doesn't have write access to the project.

Note that we still use the startsWith if condition on the job. This is because
the slash-command-action would otherwise run and fail on every comment that did
not contain a comment command. By keeping the startsWith filter the step should
either run successfully (if the commenter has write access) or fail (if they
don't).

Signed-off-by: Nic Cope <negz@rk0n.org>